assemble boxes

+++ test_python_atag/tensorflow/core/user_ops/assemble_boxes.cc
@@ -20,13 +20,13 @@ class AssembleBoxesOp : public OpKernel {
   void Compute(OpKernelContext* context) override {
     // Grab the input tensor
     const Tensor& input_tensor = context->input(0);
-    auto input = input_tensor.flat<int32>();
+    auto input = input_tensor.flat<uint16>();
 
     // Create an output tensor
     Tensor* output_tensor = NULL;
     OP_REQUIRES_OK(context, context->allocate_output(0, input_tensor.shape(),
                                                      &output_tensor));
-    auto output_flat = output_tensor->flat<int32>();
+    auto output_flat = output_tensor->flat<uint16>();
 
     // Set all but the first element of the output tensor to 0.
     const int N = input.size();
